flask_app/requirements.txt
Flask>=2.0
hdfs>=2.6.0
PyJWT>=2.0